third_party/WebKit/LayoutTests/vr/requestAnimationFrame_consistentTimestamps.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="resources/mock-vr-service.js"></script>
<canvas id="webgl-canvas"></canvas>
<script src="resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

// This test verifies that an application can hand off from window.rAF
// to vrDisplay.rAF, specifically where there is no pending vrDisplay.rAF
// at the time presentation starts.
//
// Caveat: This does *not* suffice to fully verify the current Android/GVR
// implementation. In the real implementation, the normal window VSync events
// (including window.rAF processing tied to them) are shut down, but in the
// layout test they continue running. So it's possible for this test to pass
// even though the scenario doesn't work in the real implementation. Keeping
// it as a layout test anyway as a sanity check - if this fails, we definitely
// have a problem.

vr_test( (t) => {
  return navigator.getVRDisplays().then( (displays) => {

    let display = displays[0];

    let usingPresentingRAF = false;
    let lastTime = -1;
    let lastrAFTime = -1;
    let tolerance_ms = 250;

    function onAnimationFrame(time) {
      var currentTime = window.performance.now();
      if (lastTime != -1) {
        var timeDiff = currentTime - lastTime;
        var rAFDiff = time - lastrAFTime;
        if (usingPresentingRAF) {
          t.step( () => {
            assert_less_than(Math.abs(rAFDiff - timeDiff), tolerance_ms);
          }, "timestamps not continuous");
        }
      }
      lastTime = currentTime;
      lastrAFTime = time;

      if (usingPresentingRAF) {
        t.done();
      }
      if (!display.isPresenting) {
        runWithUserGesture( () => {
          display.requestPresent([{ source : webglCanvas }]).then( () => {
            t.step( () => {
              assert_true(display.isPresenting);
            }, "Display should be presenting");
            usingPresentingRAF = true;
            display.requestAnimationFrame(onAnimationFrame);
          }, (err) => {
            t.step( () => {
              assert_unreached(err);
            }, "requestPresent rejected");
          });
        });
      }
    }
    display.requestAnimationFrame(onAnimationFrame);
  }, (err) => {
    t.step( () => {
      assert_unreached("getVRDisplays rejected");
    });
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"requestAnimationFrame timestamp is consistent between magic window and presentation.");

</script>
